tawi_project/requirements.txt
asgiref==3.5.2
Django==3.2.16
django-crispy-forms==1.14.0
Pillow==9.4.0
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.4.0
